pyproject.toml
[project]
name = "jira-github-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "openai>=1.107.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "rich>=14.1.0",
]
